+++ composeApp/src/desktopMain/kotlin/com/plusmobileapps/sample/anvilkmp/Platform.jvm.kt
@@ -1,6 +1,6 @@
 package com.plusmobileapps.sample.anvilkmp
 
-class JVMPlatform: Platform {
+class JVMPlatform : Platform {
     override val name: String = "Java ${System.getProperty("java.version")}"
 }
