+++ composeApp/src/iosMain/kotlin/com/plusmobileapps/sample/anvilkmp/Platform.ios.kt
@@ -2,8 +2,9 @@ package com.plusmobileapps.sample.anvilkmp
 
 import platform.UIKit.UIDevice
 
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+class IOSPlatform : Platform {
+    override val name: String =
+        UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
 }
 
 actual fun getPlatform(): Platform = IOSPlatform()
